finished unit testing for zipped list
