+++ javascript/linked-list/__tests__/linked-list.test.js
@@ -95,20 +95,4 @@ describe("Linked List", () => {
     list.insert(1);
     expect(list.kthFromEnd(3)).toEqual(1);
   });
-
-  it("Zips lists", () => {
-    const list1 = new LinkedList();
-    list1.insert(1);
-    list1.insert(3);
-    list1.insert(2);
-
-    const list2 = new LinkedList();
-    list2.append(5);
-    list2.append(9);
-    list2.append(4);
-
-    const zipped = zipLists(list1, list2);
-
-    expect(zipped.read()).toEqual([2, 5, 3, 9, 1, 4]);
-  });
 });
